apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.10.0
python-jose[cryptography]==3.3.0
